// lib/api/services/service/service.ts
export class Service {
    id: string = "";
    name: string = "";
    description: string = "";
    price: number = 0;
    duration: number = 0;
    category: string = "Básico";
    status: string = "Activo";
    include: string = "";
    require: string = "";
    notes: string = "";
    popularity: number = 0;
}